+++ R/demo_stim.R
@@ -12,7 +12,7 @@
 #' 
 #' The images from `demo_stim()` and `demo_unstandard()` are usable on a CC-BY license, citing: 
 #' 
-#' DeBruine, L. (2016).Young adult composite faces (Version1). figshare. doi: [10.6084/m9.figshare.4055130.v1](https://doi.org/10.6084/m9.figshare.4055130.v1). 
+#' DeBruine, L. (2016).Young adult composite faces (Version1). figshare. \doi{https://doi.org/10.6084/m9.figshare.4055130.v1} 
 #' 
 #' The image from `demo_tems()` is Lisa DeBruine (the author of webmorphR) and available on a CC-O license (no attribution needed). 
 #'
@@ -36,11 +36,13 @@ demo_stim <- function(pattern = NULL) {
 #' @rdname demo_stim
 #' 
 #' @examples
+#' \donttest{
 #' # visualise templates
 #' demo_tems() |> 
 #'   draw_tem(pt.size = 10) |> 
 #'   label() |>
 #'   plot(maxwidth = 1000)
+#' }
 demo_tems <- function(pattern = NULL) {
   path <- file.path("extdata", "tem_examples") |>
     system.file(package = "webmorphR")
@@ -53,12 +55,14 @@ demo_tems <- function(pattern = NULL) {
 #' @rdname demo_stim
 #' 
 #' @examples
+#' \donttest{
 #' # visualise keeping relative sizes
 #' demo_unstandard() |> 
 #'   to_size(keep_rels = TRUE) |>
 #'   pad(80, 0, 0, 0) |>
 #'   label() |>
 #'   plot(nrow = 2, maxwidth = 1000)
+#' }
 demo_unstandard <- function(pattern = NULL) {
   path <- file.path("extdata", "unstandard") |>
     system.file(package = "webmorphR")