+++ R/tem.R
@@ -16,12 +16,14 @@
 #' fpp106 <- tem_def("fpp106")
 #' fpp106$lines |> str()
 #'
+#' \donttest{
 #' fpp83 <- tem_def("fpp83")
 #' fpp83$mask |> str()
 #'
 #' frl <- tem_def("frl")
 #' frl$points[1:10, ]
 #' viz_tem_def(frl, pt.size = 10, line.size = 5)
+#' }
 tem_def <- function(tem_id = "frl", path = NULL) {
 
   # read file or url ----